update to 9.4.0

+++ TMessagesProj/src/main/java/org/telegram/ui/Components/StorageDiagramView.java
@@ -271,10 +271,10 @@ public class StorageDiagramView extends View implements NotificationCenter.Notif
             text1.setTextColor(Theme.getColor(Theme.key_dialogTextBlack));
             text2.setTextColor(Theme.getColor(Theme.key_dialogTextBlack));
             if (dialogId != null) {
-                int textWidth = text1.getCurrentWidth() + AndroidUtilities.dp(4) + text2.getCurrentWidth();
-                int leftpad = (getWidth() - textWidth) / 2;
-                text1.setBounds(0, AndroidUtilities.dp(115), leftpad + text1.getCurrentWidth(), AndroidUtilities.dp(115 + 30));
-                text2.setBounds(leftpad + textWidth - text2.getCurrentWidth(), AndroidUtilities.dp(115 + 3), getWidth(), AndroidUtilities.dp(115 + 3 + 30));
+                float textWidth = text1.getCurrentWidth() + AndroidUtilities.dp(4) + text2.getCurrentWidth();
+                float leftpad = (getWidth() - textWidth) / 2;
+                text1.setBounds(0, AndroidUtilities.dp(115), (int) (leftpad + text1.getCurrentWidth()), AndroidUtilities.dp(115 + 30));
+                text2.setBounds((int) (leftpad + textWidth - text2.getCurrentWidth()), AndroidUtilities.dp(115 + 3), getWidth(), AndroidUtilities.dp(115 + 3 + 30));
             }
             text1.draw(canvas);
             text2.draw(canvas);
